separate stringify from entry
